Le fasta fonctionne

+++ src/read_fasta.h
@@ -3,4 +3,4 @@
 
 
 char pick_nucleotide();
-void read_fasta(std::vector<double>& f, std::vector<std::string>& alleles, std::vector<int> marqueurs, int& N, size_t& A, std::string& file);
+void read_fasta(std::vector<double>& f, std::vector<std::string>& alleles, std::vector<size_t> marqueurs, int& N, size_t& A, std::string& file);
